Added Access-Control headers; and passwords are now checked before deleting/updating a user

+++ routes/users.js
@@ -23,34 +23,46 @@ router.post('/', (req, res) => {
       res.send(user);
     });
   } else {
-    res.status(400).send('Invalid User');
+    res.status(400).send('Invalid Username or password');
   }
 });
 
 router.put('/:username', (req, res) => {
   const username = req.body.username;
   const password = req.body.password;
+  const newPassword = req.body.newPassword;
 
-  // TODO: Check password again
-
-  if (models.User.isValidUsername(username) && models.User.isValidPassword(password)) {
-    // Generate salt and hash password
-    models.userDAO.update(new models.User(username, password), (user) => {
-      delete user.password;
-      res.send(user);
-    });
-  } else {
-    res.status(400).send('Invalid Input');
-  }
+  // Check old password
+  models.userDAO.checkCredentials(username, password, (success) => {
+    if (success) {
+      if (models.User.isValidUsername(username) && models.User.isValidPassword(newPassword)) {
+        // Generate salt and hash password
+        models.userDAO.update(new models.User(username, newPassword), (user) => {
+          delete user.password;
+          res.send(user);
+        });
+      } else {
+        res.status(400).send('Invalid Input');
+      }
+    } else {
+      res.status(401).send('Invalid credentials');
+    }
+  });
 });
 
 router.delete('/:username', (req, res) => {
   const username = req.params.username;
+  const password = req.body.password;
 
-  // TODO: Check password again
-
-  models.userDAO.delete(username, (success) => {
-    res.send(success);
+  // Check password
+  models.userDAO.checkCredentials(username, password, (success) => {
+    if (success) {
+      models.userDAO.delete(username, (success) => {
+        res.send(success);
+      });
+    } else {
+      res.status(401).send('Invalid credentials');
+    }
   });
 });
